Clone default headers to avoid foreign mutation
Maps are pass-by-reference, so multiple servers would have all used the
same map reference, potentially leading to racy mutability bugs.

Amusingly, this commit fixes a mutability bug that relied on prior calls
setting the X-Plex-Token in the global `headers` object for later calls
to work. This fixes that by reusing the server specific headers with the
correct token already set.

Signed-off-by: Joe Groocock <me@frebib.net>